Refactor ConstantExpression to fix SQ warnings
