+++ UnitTestProject1/SimpleStatementTests.cs
@@ -7,6 +7,7 @@ using Oberon0.CompilerSupport;
 
 namespace Oberon0.Compiler.Tests
 {
+    using Oberon0.Compiler.Expressions.Constant;
 
     [TestFixture]
     public class SimpleStatementTests
