+++ oberon0/Definitions/Module.StandardFunctions.cs
@@ -6,6 +6,8 @@ using Oberon0.Compiler.Types;
 
 namespace Oberon0.Compiler.Definitions
 {
+    using Oberon0.Compiler.Expressions.Constant;
+
     public partial class Module
     {
         private void DeclareStandardFunctions()
