+++ oberon0/Expressions/Expression.cs
@@ -7,6 +7,15 @@ namespace Oberon0.Compiler.Expressions
 {
     public abstract class Expression
     {
+        protected Expression()
+        {
+        }
+
+        protected Expression(TypeDefinition targetType)
+        {
+            TargetType = targetType;
+        }
+
         public int Operator { get; set; }
 
         public TypeDefinition TargetType { get; set; }
