+++ oberon0/Expressions/Operations/OpRelop.cs
@@ -17,6 +17,7 @@ namespace Oberon0.Compiler.Expressions.Operations
     using JetBrains.Annotations;
 
     using Oberon0.Compiler.Definitions;
+    using Oberon0.Compiler.Expressions.Constant;
     using Oberon0.Compiler.Expressions.Operations.Internal;
     using Oberon0.Compiler.Types;
