+++ oberon0/Expressions/Operations/OpSubNumber.cs
@@ -15,6 +15,7 @@ namespace Oberon0.Compiler.Expressions.Operations
     using JetBrains.Annotations;
 
     using Oberon0.Compiler.Definitions;
+    using Oberon0.Compiler.Expressions.Constant;
     using Oberon0.Compiler.Expressions.Operations.Internal;
     using Oberon0.Compiler.Types;
